refactor: move shared test utilities to dedicated module
Move common utilities used by i18n test files into utils.ts:
- File operations (fileExists, loadFileContent)
- JSON parsing and path handling
- Logging utilities

Replace direct logBuffer access with clearLogs() function
Remove duplicate code between test files
